Coin Flipper
The coin flip game is a simple web application created using HTML, CSS, and JavaScript. The user interface consists of a button styled as a coin, and upon clicking it, a flip animation is triggered using CSS transitions and keyframes. The animation simulates the motion of a coin flipping. And JavaScript is used to dynamically add and remove outcome whether it's a head or tail and prints the result.